Fix file scheme URL parsing for Firefox
The URL specification (https://url.spec.whatwg.org/#concept-url-origin)
says, regarding the origin for file scheme:

Unfortunate as it is, this is left as an exercise to the reader. When in doubt, return a new opaque origin.

The opaque origin serializes as "null".

Mozilla developer says
(https://developer.mozilla.org/en-US/docs/Web/API/URL/origin), of the
origin property of URL:

For URL using file: scheme, the value is browser dependent.

The original tests were run against Chrome browser and expected origin
to be 'file://' but Firefox returns null.

This change tests for the file scheme and returns 'file://' as origin in
this case, regardless of what the browser's URL object returns.